move gce cloud provider initialization to main
